refactor: Move sourcer pkg to pkg
