+++ cmd/convert.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/glass-cms/glasscms/internal/item"
 	"github.com/glass-cms/glasscms/parser"
-	"github.com/glass-cms/glasscms/sourcer"
+	"github.com/glass-cms/glasscms/pkg/sourcer"
 	"github.com/lmittmann/tint"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
