+++ pkg/sourcer/file_system_test.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/glass-cms/glasscms/sourcer"
+	"github.com/glass-cms/glasscms/pkg/sourcer"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
